// src/app/interfaces/orderdetails.ts
export interface Orderdetails {
    // orderId : number;
    // userId : number;
    // intervalId : number;
    // typeId : number;
    // durationId : number;
    // orderPlaceDate : Date;
    // orderDate : Date;
    // areaId : number;
    // address : string;
    // paymentId : string;
    // totalCharge : number;
    // restaurantsId : number;
    // intervalName : string;
    // typeName : string;
    // durationTime : string;
    // areaName : string;
    // restaurantName : string;

    address: string;
    areaId: number;
    areaName:  string;
    durationId: number;
    durationTime:  string;
    intervalId: number;
    intervalName:  string;
    orderDate: Date;
    orderId: number;
    orderPlaceDate: Date;
    paymentId:  string;
    restaurantName:  string;
    restaurantsId: number;
    totalCharge: number;
    typeId: number;
    typeName:  string;
    userId: number;
}